requirements.txt
click==8.1.7
gitdb==4.0.12
GitPython==3.1.43
markdown-it-py==4.0.0
mdurl==0.1.2
neo4j==5.19.0
Pygments==2.19.2
python-dotenv==1.0.1
pytz==2025.2
rich==13.7.1
smmap==5.0.2
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
